requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
python-dotenv>=1.0.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
openai>=1.0.0
langgraph>=0.1.0
langchain-core>=0.2.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
httpx>=0.27.0
crawl4ai>=0.4.0

